+++ marshmallow/exceptions.py
@@ -20,7 +20,7 @@ class ValidationError(MarshmallowError):
     :param list fields: `Field` objects to which the error applies.
     """
 
-    def __init__(self, message, field_names=None, fields=None, data=None):
+    def __init__(self, message, field_names=None, fields=None, data=None, **kwargs):
         if not isinstance(message, dict) and not isinstance(message, list):
             messages = [message]
         else:
@@ -38,6 +38,7 @@ class ValidationError(MarshmallowError):
             self.field_names = field_names or []
         # Store nested data
         self.data = data
+        self.kwargs = kwargs
         MarshmallowError.__init__(self, message)